+++ src/protocol/protocol.h
@@ -83,7 +83,7 @@ struct diagpkt_subsys_hdr {
 	uint16_t	subsys_cmd_code;
 } __attribute__((packed));
 
-int diag_push_subsys_hdr(struct msgb *msg, uint8_t subsys, uint8_t code);
+int diag_push_subsys_hdr(struct msgb *msg, uint8_t subsys, uint16_t code);
 uint32_t diag_ts_to_fn(uint64_t ts);
 uint32_t diag_ts_to_epoch(uint64_t qd_time);
